Fix 2025 decimal cutoff rank corruption in database schemas and import scripts

+++ backend/schema_postgres.sql
@@ -77,7 +77,7 @@ CREATE TABLE IF NOT EXISTS cutoffs (
     course_id INTEGER NOT NULL,
     round INTEGER NOT NULL,
     category TEXT NOT NULL,
-    cutoff_rank BIGINT NOT NULL,
+    cutoff_rank REAL NOT NULL,
     year INTEGER DEFAULT 2025,
     FOREIGN KEY(course_id) REFERENCES courses(id) ON DELETE CASCADE
 );
